frontend_app/attached_assets/requirements.txt
Flask
pymongo
pydantic
python-dotenv
